+++ tests/util/test_decorators.py
@@ -1,34 +1,9 @@
 # SPDX-License-Identifier: BSD-2-Clause
 
-import warnings
-
 from torii.test import ToriiTestCase
 from torii.util import decorators
 
 class DecoratorsTestCase(ToriiTestCase):
-
-	def test_memoize(self):
-		hittest: int = 0
-
-		with warnings.catch_warnings():
-			# Ignore the new Deprecation warning
-			warnings.simplefilter('ignore')
-
-			@decorators.memoize
-			def foo(a: int, b: int) -> int:
-				nonlocal hittest
-				hittest = hittest + 1
-				return a * 2 + b
-
-		self.assertEqual(hittest, 0)
-		_ = foo(1, 2)
-		self.assertEqual(hittest, 1)
-		_ = foo(1, 2)
-		self.assertEqual(hittest, 1)
-		_ = foo(2, 1)
-		self.assertEqual(hittest, 2)
-		_ = foo(2, 1)
-		self.assertEqual(hittest, 2)
 
 	def test_final(self):